+++ lib/PixelBufferLib/src/bufferTypes/LogoPixelBuffer.cpp
@@ -5,8 +5,8 @@ LogoPixelBuffer::LogoPixelBuffer(int gpioPin) : PixelBuffer_Old(gpioPin)
     // 140 actual pixels
     // 23 columns, 24 rows
     m_numPixels = 140;
-    m_pixelBufferSize = std::max((uint) PB_MINIMUM_PIXELS, m_numPixels);
-    m_pixelColors = new uint32_t[m_pixelBufferSize];
+    m_pixelBufferSize = std::max((unsigned int) PB_MINIMUM_PIXELS, m_numPixels);
+    m_pixelColors = new unsigned long[m_pixelBufferSize];
 
     m_columns.push_back(new std::vector<int>{ 99, 100});
     m_columns.push_back(new std::vector<int>{ 98, 102});
@@ -59,7 +59,7 @@ LogoPixelBuffer::LogoPixelBuffer(int gpioPin) : PixelBuffer_Old(gpioPin)
 
     m_digits.push_back(new std::vector<int>());
     
-    for (uint i = 0; i < m_numPixels; i++)
+    for (int i = 0; i < m_numPixels; i++)
     {
         m_digits[0]->push_back(i);
     }
